pyproject.toml
[project]
name = "pulse-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "pydantic[email]>=2.12.5",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.1",
    "uvicorn>=0.40.0",
]
